backend/pyproject.toml
[project]
name = "agentic-learning-backend"
version = "0.1.0"
description = "Google ADK agents and tools for the agentic learning app"
# Keep in sync with /.python-version
requires-python = "==3.13.*"
dependencies = [
    "google-adk>=1.0.0",
    "psycopg2-binary>=2.9",
    "google-cloud-firestore>=2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.0",
    "google-cloud-secret-manager>=2.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "ruff>=0.15.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
